requirements.txt
python-dotenv >= "1.0.1"
langchain >= "0.2.1"
langchain-openai >= "0.1.8"